requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
SQLAlchemy==1.4.25
Werkzeug==2.0.2
DAWG-Python==0.7.2
docopt==0.6.2
intervaltree==3.1.0
ipymarkup==0.9.0
natasha==1.4.0
navec==0.10.0
numpy==1.21.3
pdf2image==1.16.0
Pillow==8.4.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytesseract==0.3.8
razdel==0.5.0
slovnet==0.5.0
sortedcontainers==2.4.0
tesseract==0.1.3
yargy==0.15.0
